Solutions/2017/AdventOfCode2017.Tests/Day07/Day07SolverTests.cs
namespace AdventOfCode2017.Tests.Day07
{
    using System.Collections.Generic;
    using AdventOfCode2017.Day07;
    using Xunit;

    public class Day07SolverTests
    {
        private readonly Day07Solver subject = new Day07Solver();

        private static string TestData =
@"pbga (66)
xhth (57)
ebii (61)
havc (66)
ktlj (57)
fwft (72) -> ktlj, cntj, xhth
qoyq (66)
padx (45) -> pbga, havc, qoyq
tknk (41) -> ugml, padx, fwft
jptl (61)
ugml (68) -> gyxo, ebii, jptl
gyxo (61)
cntj (57)";

        public static IEnumerable<object[]> PuzzleOneData = new[]
        {
            new object[] { "tknk", TestData }
        };

        [Theory]
        [MemberData(nameof(PuzzleOneData))]
        public void PuzzleOne(string expected, string input)
        {
            var actual = this.subject.PuzzleOne(input);

            Assert.Equal(expected, actual);
        }

        public static IEnumerable<object[]> PuzzleTwoData = new[]
        {
            new object[] { 60, TestData }
        };

        [Theory]
        [MemberData(nameof(PuzzleTwoData))]
        public void PuzzleTwo(int expected, string input)
        {
            var actual = this.subject.PuzzleTwo(input);

            Assert.Equal(expected, actual);
        }
    }
}